extract AuthenticateForm
