+++ App/Components/AccountForm/SwitchType/index.js
@@ -8,6 +8,6 @@ import type {AccountType} from '../index';
 export const SwitchType = ({onPress, type}: {type: AccountType, onPress: () => void}) => (
   <Text style={styles.container} onPress={onPress}>
     {type === 'createUser' ?
-      '既にアカウントをお持ちの方はこちらをタップ' : 'まだアカウントをお持ちでない方はこちらをタップ'}
+      '既にアカウントをお持ちの方はこちら' : 'まだアカウントをお持ちでない方はこちら'}
   </Text>
 );
