+++ Tests/mock/redux/index.js
@@ -1,7 +1,8 @@
 // @flow
-import type {TDefaultChannels, TDefaultUser} from '../../../App/types/';
+import type {TDefaultChannels, TDefaultUser, TUserActions} from '../../../App/types/';
 import {channelsStoreWithKeyMock} from '../channel';
 import {userMock} from '../user';
+import {noop} from '../index';
 
 export const defaultChannelsMock: TDefaultChannels = {
   items: channelsStoreWithKeyMock(),
@@ -13,6 +14,25 @@ export const defaultChannelsMock: TDefaultChannels = {
 
 export const defaultUserMock: TDefaultUser = {
   item: userMock,
+  isForgotPassword: false,
   isFetching: false,
   errorMessage: '',
+};
+
+export const userActionsMock: TUserActions = {
+  userRequest: noop,
+  userLogin: (email: string, password: string) => {
+    console.log('userActionsMock.userLogin');
+    console.log(email, password);
+  },
+  userCreate: (email: string, password: string) => {
+    console.log('userActionsMock.userCreate');
+    console.log(email, password);
+  },
+  userSuccess: noop,
+  userFailure: noop,
+  userUpdateProfile: noop,
+  userSendEmailVerification: noop,
+  userReload: noop,
+  userSwitchForgotPassword: noop,
 };
